requirements.txt
# Existing dependencies (keep what you already have)
fastapi
uvicorn
sqlalchemy
psycopg2-binary  # or psycopg2 if you have it
python-jose[cryptography]
passlib[bcrypt]
python-multipart
pydantic[email]
google-auth
python-dotenv

# New dependencies for Apple Sign In
pyjwt
cryptography
requests

#payment 
stripe

# OpenAI for Whisper STT
openai>=1.10.0

# Audio Processing
numpy>=1.24.0
scipy>=1.11.0

# HTTP Client (for async requests)
httpx>=0.26.0